First code commit
This commit lays down the basic design of the code. It can do little of
interest (aside from take a JSON file and print to stdout), but it makes
clear how the program hangs together.

To understand it, follow the comments, check out the class diagram in
the doc folder and look at the test JSON file. Also, a tiny example
Visitor called SillyVisitor is in the doc folder.

Change-Id: Id3f1c0f5e54577888c2fc67d2d2d171306c4ff4a