Fixes bug with tranche tables array index
This is a symptom of the funny way tranche tables work. The upper bound
on the last tranche table should be limitless or infinity (so that the
tranche table covers all possibilities). In practice, it is just some
really high number. If you put a low number as that upper bound, then
if the calculated figure (e.g. cumulative production) is high it will
fit nowhere in the table.

In any case, I've put a sanity check in all the loops to prevent this.

Change-Id: I239ef6c46aa6576d9f7e84f5d587470965d6e980
Ref: #954